organize/split arrow code into multiple files
