+++ src/include/duckdb/common/arrow/appender/list.hpp
@@ -1,0 +1,8 @@
+#include "duckdb/common/arrow/appender/bool_data.hpp"
+#include "duckdb/common/arrow/appender/enum_data.hpp"
+#include "duckdb/common/arrow/appender/list_data.hpp"
+#include "duckdb/common/arrow/appender/map_data.hpp"
+#include "duckdb/common/arrow/appender/scalar_data.hpp"
+#include "duckdb/common/arrow/appender/struct_data.hpp"
+#include "duckdb/common/arrow/appender/union_data.hpp"
+#include "duckdb/common/arrow/appender/varchar_data.hpp"
