+++ src/include/duckdb/common/arrow/arrow_appender.hpp
@@ -27,6 +27,11 @@ public:
 	//! Returns the underlying arrow array
 	DUCKDB_API ArrowArray Finalize();
 
+public:
+	static void ReleaseArray(ArrowArray *array);
+	static ArrowArray *FinalizeChild(const LogicalType &type, ArrowAppendData &append_data);
+	static unique_ptr<ArrowAppendData> InitializeChild(const LogicalType &type, idx_t capacity, ArrowOptions &options);
+
 private:
 	//! The types of the chunks that will be appended in
 	vector<LogicalType> types;
